+++ docs/CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 -----
 - Add  ``REDIS_CONNECTION_STRING`` setting needed by the ``Resolwe``
 - Add process ``samtools-view``
+- Add process ``samtools-coverage``
 
 Changed
 -------